\name{MOL.plot}
\alias{MOL.plot}
\title{
Quick Plots for DiffusionRimp Objects
}
\description{
\code{MOL.plot()} recognizes output objects calculated using routines from the \bold{DiffusionRimp} package and subsequently constructs an appropriate plot, for example a perspective plot of a transition density.
}
\usage{
MOL.plot(x)
}
\arguments{
  \item{x}{
Generic MOL-objects, i.e. \code{res = MOL.density()}.
}
}

\value{Varies in accordance with input type.
}
\author{
Etienne A.D. Pienaar: \email{etiannead@gmail.com}
}
\references{
Updates available on GitHub at \url{https://github.com/eta21}.
}


\seealso{
\code{\link{MOL.density}}, \code{\link{BiMOL.density}}, \code{\link{MOL.passage}} etc.
}
\examples{
  \donttest{
#===============================================================================
  }
}
\keyword{plot}

