\name{DiffusionRimp-package}
\alias{DiffusionRimp-package}
\alias{DiffusionRimp}
\docType{package}
\title{
Inference for diffusion processes by imputing missing data with Brownian bridges.
}
\description{
A package for performing data imputation on discretely observed diffusion processes. 
}
\details{
\tabular{ll}{
Package: \tab DiffusionRimp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-06\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Etienne A.D. Pienaar \email{etiennead@gmail.com}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
example(RSimpute.mcmc)
}
